requirements.txt
google-api-python-client>=2.160.0
google-auth>=2.38.0
google-auth-oauthlib>=1.2.1
python-dotenv>=1.0.1
openai>=1.61.0
